requirements.txt
ddgs
python-dotenv>=1.0
google-genai>=0.3
requests>=2.31
line-bot-sdk>=3.0.0